+++ image/service/mmc-agent/assets/get-config.sh
@@ -1,5 +1,9 @@
 #!/bin/bash -e
 
+# set -x (bash debug) if log level is trace
+# https://github.com/osixia/docker-light-baseimage/blob/stable/image/tool/log-helper
+log-helper level eq trace && set -x
+
 # get mmc plugin config
 # usage : get-config.sh plugin_variable config