+++ image/service/mmc-agent/install.sh
@@ -4,7 +4,7 @@
 dpkg -i /container/service/mmc-agent/assets/package/python-mmc-core_3.9.90-10_all.deb
 dpkg -i /container/service/mmc-agent/assets/package/python-mmc-base_3.9.90-10_all.deb
 dpkg -i /container/service/mmc-agent/assets/package/python-mmc-dashboard_3.9.90-10_all.deb
-dpkg -i /container/service/mmc-agent/assets/package/mmc-agent_3.9.90-10_all.deb
+dpkg -i /container/service/mmc-agent/assets/package/mmc-agent_3.9.90-10_all.deb || true
 
 rm -rf /container/service/mmc-agent/assets/package