+++ image/service/mmc-agent/process.sh
@@ -4,4 +4,7 @@
 # https://github.com/osixia/docker-light-baseimage/blob/stable/image/tool/log-helper
 log-helper level eq trace && set -x
 
-exec /usr/sbin/mmc-agent -d
+ARG="-s" # Do not daemonize and not log debug
+log-helper level gt info && ARG="-d" # Do not daemonize
+
+exec /usr/sbin/mmc-agent ${ARG}