+++ image/service/mmc-agent/startup.sh
@@ -50,6 +50,9 @@ if [ ! -e "$FIRST_START_DONE" ]; then
   # base plugin configuration
   ${CONTAINER_SERVICE_DIR}/mmc-agent/assets/config-plugin.sh "MMC_AGENT_BASE_PLUGIN_CONFIG" /etc/mmc/plugins/base.ini
 
+  # dashboard plugin configuration
+  ${CONTAINER_SERVICE_DIR}/mmc-agent/assets/config-plugin.sh "MMC_AGENT_DASHBOARD_PLUGIN_CONFIG" /etc/mmc/plugins/dashboard.ini
+
   cp -f /etc/mmc/agent/config.ini ${CONTAINER_SERVICE_DIR}/mmc-agent/assets/config.ini
   cp -f /etc/mmc/plugins/base.ini ${CONTAINER_SERVICE_DIR}/mmc-agent/assets/base.ini